+++ lib/src/chart/YieldCurveChart.tsx
@@ -34,7 +34,11 @@ const YieldCurveChartRenderFunction: ForwardRefRenderFunction<
   );
 
   return (
-    <div ref={containerRef} {...containerProps}>
+    <div
+      ref={containerRef}
+      aria-hidden={containerProps?.["aria-hidden"] ?? true}
+      {...containerProps}
+    >
       {!!container && (
         <ChartComponent<number, IYieldCurveChartApi, YieldCurveChartOptions>
           container={container}