+++ lib/src/primitives/types.ts
@@ -1,5 +1,7 @@
 import type {
   IChartApiBase,
+  IPaneApi,
+  IPanePrimitive,
   ISeriesApi,
   ISeriesPrimitive,
   SeriesType,
@@ -39,6 +41,39 @@ export type RenderPrimitive<T extends SeriesType = SeriesType, HorzScaleItem = T
   series: ISeriesApi<T, HorzScaleItem>;
 }) => ISeriesPrimitive<HorzScaleItem>;
 
+/**
+ * Pane primitive API reference type that can be used to access the pane primitive plugin API.
+ */
+export type PanePrimitiveApiRef<HorzScaleItem = Time> = {
+  /**
+   * Internal reference to the pane primitive API instance.
+   */
+  _primitive: IPanePrimitive<HorzScaleItem> | null;
+  /**
+   * Function to get the pane primitive API instance.
+   */
+  api(): IPanePrimitive<HorzScaleItem> | null;
+  /**
+   * Function to initialize the pane primitive API instance.
+   */
+  init(): IPanePrimitive<HorzScaleItem> | null;
+  /**
+   * Function to clear the pane primitive API instance.
+   */
+  clear(): void;
+};
+
+/**
+ * Render function type for pane primitives.
+ */
+export type RenderPanePrimitive<HorzScaleItem = Time> = ({
+  chart,
+  pane,
+}: {
+  chart: IChartApiBase<HorzScaleItem>;
+  pane: IPaneApi<HorzScaleItem>;
+}) => IPanePrimitive<HorzScaleItem>;
+
 type SeriesPrimitivePropsWithRender<T extends SeriesType, HorzScaleItem = Time> = {
   render: RenderPrimitive<T, HorzScaleItem>;
   plugin?: never;
@@ -46,6 +81,16 @@ type SeriesPrimitivePropsWithRender<T extends SeriesType, HorzScaleItem = Time> 
 
 type SeriesPrimitivePropsWithPlugin<HorzScaleItem = Time> = {
   plugin: ISeriesPrimitive<HorzScaleItem>;
+  render?: never;
+};
+
+type PanePrimitivePropsWithRender<HorzScaleItem = Time> = {
+  render: RenderPanePrimitive<HorzScaleItem>;
+  plugin?: never;
+};
+
+type PanePrimitivePropsWithPlugin<HorzScaleItem = Time> = {
+  plugin: IPanePrimitive<HorzScaleItem>;
   render?: never;
 };
 
@@ -58,3 +103,10 @@ export type SeriesPrimitiveProps<
 > =
   | SeriesPrimitivePropsWithRender<T, HorzScaleItem>
   | SeriesPrimitivePropsWithPlugin<HorzScaleItem>;
+
+/**
+ * Pane primitive properties that can be used to create a pane primitive.
+ */
+export type PanePrimitiveProps<HorzScaleItem = Time> =
+  | PanePrimitivePropsWithRender<HorzScaleItem>
+  | PanePrimitivePropsWithPlugin<HorzScaleItem>;